Broke down data_structures.py and parts of other parsing files into more of a single responsibility files and into directories
