+++ Part2/Part2_Backend/parsing/file_utils.py
@@ -1,72 +1,9 @@
 import pandas as pd
-from io import BytesIO
-from data_structures import CoordinateFile, MatrixFile, FileProcessingConfig
+from core.coordinate_file import CoordinateFile
+from core.matrix_file import MatrixFile
+from core.config import FileProcessingConfig
+from parsing.domain_utils import process_domain_field
 
-def validate_file_extension(filename: str, file_type: str) -> None:
-    valid_extensions = {
-        'matrix': ['.xlsx', '.csv', '.tsv'],
-        'coordinate': ['.xlsx', '.csv', '.tsv']
-    }
-    
-    if not any(filename.lower().endswith(ext) for ext in valid_extensions[file_type]):
-        raise ValueError(
-            f"Invalid {file_type} file format. Supported formats are: " + 
-            ", ".join(ext.upper() for ext in valid_extensions[file_type])
-        )
-
-def clean_dataframe_whitespace(df: pd.DataFrame) -> pd.DataFrame:
-    """Clean whitespace from DataFrame index, columns, and string values."""
-    # Clean index and column names
-    if df.index.dtype == "object":
-        df.index = df.index.str.strip()
-    df.columns = df.columns.str.strip()
-    # Clean string values in all columns
-    for col in df.columns:
-        if df[col].dtype == "object":
-            # Convert to string first, then strip
-            df[col] = df[col].astype(str).str.strip()
-            # Convert back to original type if possible
-            try:
-                if df[col].str.isnumeric().all():
-                    df[col] = pd.to_numeric(df[col])
-            except:
-                pass  # Keep as string if conversion fails
-    
-    return df
-
-def read_file(file, file_type: str) -> pd.DataFrame:
-    """Read a file into a pandas DataFrame based on its extension."""
-    # Handle both file objects and BytesIO objects
-    if hasattr(file, 'name'):
-        filename = file.name.lower()
-        validate_file_extension(filename, file_type)
-    else:
-        # For BytesIO objects, we need to determine the file type from the content
-        # For now, we'll assume it's an Excel file since that's the most common case
-        filename = 'temp.xlsx'
-    
-    try:
-        if filename.endswith('.csv'):
-            df = pd.read_csv(BytesIO(file.read()), encoding='utf-8')
-        elif filename.endswith('.tsv'):
-            df = pd.read_csv(BytesIO(file.read()), sep='\t', encoding='utf-8')
-        else:  # Excel file
-            df = pd.read_excel(BytesIO(file.read()), engine='openpyxl')
-            
-        # Clean whitespace from the DataFrame
-        df = clean_dataframe_whitespace(df)
-            
-    except UnicodeDecodeError:
-        raise ValueError("File encoding error. Please ensure the file is UTF-8 encoded.")
-    except pd.errors.ParserError as e:
-        if filename.endswith('.csv'):
-            raise ValueError("CSV parsing error. Please ensure the file is properly formatted with comma separators.")
-        elif filename.endswith('.tsv'):
-            raise ValueError("TSV parsing error. Please ensure the file is properly formatted with tab separators.")
-        else:
-            raise ValueError(f"Excel parsing error: {str(e)}")
-    
-    return df
 
 def validate_matrix_coordinate_mapping(matrix_df: pd.DataFrame, coord_df: pd.DataFrame) -> None:
     """Validate that all matrix indices exist in the coordinate file's name column."""
@@ -153,7 +90,6 @@ def parse_coordinates(coord_file, include_domains=False):
         
         if include_domains:
             # Process domain fields (this is specific to domain parsing)
-            from domain_parse import process_domain_field
             domain_names, domain_col_names = process_domain_field(cleaned_data)
             
             # Return only the required columns in the specified order
